dep: fixed pytorch poetry resolving
Finally got poetry to play nice by specifying the exact URL for the
wheels, for both torch and torchvision with CUDA 11.1

For redundancy, I've also specified the versions in the conda
environment; if poetry doesn't work, then conda it. Also, noticed that
it takes a hot minute to resolve with poetry.